moved mode switch buttons into own component
