+++ core/src/components/navbar/Tab.vue
@@ -1,23 +1,19 @@
 <template>
   <div class="navbarTab" v-if="id == activeNavbarTab">
-    <md-button class="navbarButton smallNavbarButton" v-on:click="switchEditingMode('editing')" :class="{'md-raised': editingMode == 'editing'}">
-      <md-icon>edit</md-icon>
-    </md-button>
-    <md-button class="navbarButton smallNavbarButton" v-on:click="switchEditingMode('selecting')" :class="{'md-raised': editingMode == 'selecting'}">
-      <md-icon>highlight_alt</md-icon>
-    </md-button>
-    <md-button class="navbarButton smallNavbarButton" v-on:click="switchEditingMode('drawing')" :class="{'md-raised': editingMode == 'drawing'}">
-      <md-icon>gesture</md-icon>
-    </md-button>
-    <md-button class="navbarButton smallNavbarButton" />
+    <mode-switch-buttons />
     <slot />
   </div>
 </template>
 
 <script>
+import modeSwitchButtons from "./ModeSwitchButtons.vue";
+
 import { mapState } from "vuex";
 
 export default {
+  components: {
+    modeSwitchButtons,
+  },
 	props: {
 		"id": {
 			default: 0,
@@ -25,13 +21,7 @@ export default {
 	},
 	computed: mapState({
 		activeNavbarTab: state => state.core.activeNavbarTab,
-		editingMode: state => state.core.editingMode,
 	}),
-	methods: {
-		switchEditingMode: function(mode) {
-			this.$store.commit("switchEditingMode", {mode, }, {module: "core" });
-		},
-	},
 };
 </script>
